Dockerfile: Optimize image layers to enable better caching
Update the root Dockerfile and replace the generic `COPY . .` command
with only the files and directories needed to create the provisioner
binar(ies). This should help speed up builds as we're not invalidating
the image layer when we modify a file that doesn't affect the actual
build.

Also update the Dockerfile and ensure we're downloading Go
modules at the beginning of the build stage to help cache dependencies
and decrease over build times over subsequent builds.

Signed-off-by: timflannagan <timflannagan@gmail.com>