kuberpak: Remove the controllers/suite_test.go as it relies on envtest
Remove the suite_test.go testing file as our `make test-unit` target
just calls `go test ...` under-the-hood and we don't necessarily need to
use `envtest` to test the current set of provisioners for the time
being. Update the go.mod and drop the ginkgo and gomega libraries as
direct dependencies.

Signed-off-by: timflannagan <timflannagan@gmail.com>