Makefile,Dockerfile: Fix the runtime errors in the rukpak container image
Update the Makefile and ensure that cgo is disabled for both the
registry+v1 and plain bundle provisioners. Update the Dockerfile to
specify an explicit path to the `k8s` executable to silence the "exec
user process caused: no such file or directory" error message when
running this container image locally.

Signed-off-by: timflannagan <timflannagan@gmail.com>