Use locally-built bundle images for e2e
This commits adds a testdata directory containing two bundles
that match the content of the existing bundles we use during the e2e
test.

It also adds a make target to build and load these bundles during e2e
runs.

The end result is that we can now more easily change the content of
the bundles being used in e2e tests alongside any changes to our code
or the tests themselves.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>