Propose a spec (for now just via code) for bundle metadata
Add machinery to parse the bundle version from this file in the
plain provisioner with special handling to assume bundles missing this
file are plain v0 bundles (at least from the perspective of the plain
provisioner).

Lastly, add specific handling for supported and unsupported bundle
formats in the plain provisioner, and add a test to ensure the plain
provisioner correctly handles unsupported bundles.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>