git bundle source: avoid copying full contents of unpack container
I was debugging the git source flake and I noticed that the log output
for the unpack pod was unexpectedly long. Turns out, the "contents" JSON
included the full filesystem of that container. All we actually need is
the bundle directory.

This commit resolves this with a few tweaks:
- Change the volume mount from "/manifests" to "/bundle".
- Just before copying, delete the repo's ".git" directory. This avoids
  copying that directory into /bundle when we're treating the entire git
  repository as the bundle.
- Copy the entire specified directory within the git repo into the
  /bundle directory (this results in "/bundle/manifests" for valid
  bundles.
- The unpack binary unpacks the "/bundle" directory rather than "/"

I was debugging the git source flake and I noticed that the log output
for the unpack pod was unexpectedly long. Turns out, the "contents" JSON
included the full filesystem of that container, which includes a few large
binaries. All we actually need is the bundle directory.

This commit resolves this with a few tweaks:
- Change the volume mount from "/manifests" to "/bundle".
- Just before copying, delete the repo's ".git" directory. This avoids
  copying that directory into /bundle when we're treating the entire git
  repository as the bundle.
- Copy the entire specified directory within the git repo into the
  /bundle directory (this results in "/bundle/manifests" for valid
  bundles.
- The unpack binary unpacks the "/bundle" directory rather than "/"

I'm not sure if this is the root cause of the flake we've been seeing,
but this fix _drastically_ decreases the size of the logs produced by
the unpack bundle, so it seems feasible and is a good change to make
regardless.