(bundle controller) Refactor to carve out phase transition
The current Reconciliation method contains phase transition logic, with the
assumption that the only componenet involved in transitioning the phase of
the Bundle object is updating the status.Phase field. That assumption does
not leave room for additional components to be added that may be needed in
the future (eg emit prometheus metric as part of phase transition). This
commit creates a logical unit for phase transition, that includes all the
current steps taken for the transition, while leaving room for additional
transition steps to be added.