remove bundle.status.info field
This field is not generic enough to cover all potential bundle formats.
For example, a helm chart does not contain specific `client.Object`s.
Rather, it contains a helm chart directory containing metadata, default
values, and some templates, which can be _rendered_ into manifests.

Thus, we should remove this field. If we later determine that there are
other fields that _do_ apply to all bundles, we can add them in
follow-on pull requests.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>