Add markdown linter to sanity github workflow
Introduce a new linter as part of the sanity check that focuses
only on linting markdown files. Having a markdown linter ensures uniform
documentation formatting and leads to higher quality documentation.

The super-linter action introduced seems like the most up-to-date and
well maintained markdown linter available as a GitHub Action.