Add bundle-template-hash label to dynamic Bundles
Adds a unique hash label to the metadata of embedded Bundles upon Bundle
creation. Storing the hash of the embedded bundle template in a label on the
Bundle results in an easier comparison when comparing an existing Bundle versus
a desired embedded Bundle. A similar pattern is used in the replicaset
controller.

This change also appends the template hash to the name of the embedded
Bundle, which results in the Bundle name being deterministic.