Move the FS and ObjectFile types to separate packages
The bundle library currently assumes that all bundles will contain or
have the possiblity to contain Kubernetes objects. This is not the case
as some bundle types, Helm bundles for example, will not contain any
manifest files. This commit moves the FS and ObjectFile types to a
separte package so bundle implementations can freely choose whether or
not it contains manifest files.