Fix bundle conversion so it creates a new MapFS
The current conversion implemetation attempts to convert bundles in
place. For example, if you pass it a registry+v1 bundle to convert to
plain+v0, the logic attempts to use the same FS and add the new
manifests to it. This does not work fully since the golang FS
abstraction does not support write operations. This commit fixes the
conversion logic so it creates a new MapFS and adds the files from the
source bundle that it needs along with any new ones. This makes it so
bundles are immutable since the MapFS is a conversion implementation
detail, not part of the public API like StoreObjects was previously.