Add ConfigMapSyncer controller
The ConfigMapSyncer syncs secret data to configmaps based on injection
annotations present in configmaps in watched namespaces.

We include a rukpak-ca configmap with these annotations present so
that cluster administrators can share rukpak-ca trust without
exposing the CA key that's present in the rukpak-ca secret.

This commit also updates the rukpakctl binary to use the configmap
rather than the secret to load the rukpak CA. This is helpful for rukpak
users that might have access to read configmaps but not secrets in the
rukpak system namespace.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>