Provisioner class names must be DNS 1123 labels
In order to ensure that we can introduce a `ProvisionerClass` API in the
future without causing breaking changes in the `Bundle` and
`BundleDeployment` APIs, we need to ensure that `Bundle` and
`BundleDeployment` objects use provisioner class name values that can
reference the `metadata.name` field of potential `ProvisionerClass`
objects. Currently, the provisioner class names use both `.` and `/`
characters, which are invalid in certain core Kubernetes resource names.

Therefore, this commit changes the plain and registry provisioner IDs to
`core-rukpak-io-plain` and `core-rukpak-io-registry` and it changes the
validation of the `Bundle` and `BundleDeployment` provisioner class name
fields to ensure only DNS-1123 label names can be used.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>