More upload source improvements
When the port forwarder fails, it results in the tar writer process
hanging indefinitely. This fixes the issue by ensuring the the writer is
closed by the time the upload process returns.

This also improves the reliability of the port forward setup by waiting
(if ncessary) for a service endpoint to be available.