Refactor source/git and add unit tests
The git.Unpack called into the git library too much to be reasonably
unit testable. Some functions broken out to improve coverage.

To test the ssh cert usage, we needed a real private key, which I
generated for this purpose. We need .gitleaks.toml to identify this
private key as safe to upload.

Signed-off-by: Austin Macdonald <austin@redhat.com>