image source: use job for bundle unpacking
It turns out that directly managing pods for image unpacking is more
susceptible to runtime mutation issues (since pods are more likely to
have mutation admission webhooks associated with them than jobs).
Mutating webhooks are problematic for our controller because we need to
reliably detect when a diff between the desired pod and the existing
on-cluster pod is a result of a change triggered via our controller vs.
a change due to some unrelated mutating webhook.

The problem still exists with a Job if the cluster has a mutating
webhook for jobs, but we'll assume for now that the problem is much less
pronounced with jobs than with pods.

NOTE: the pod specs in both jobs and pods are immutable. Therefore, our
controller cannot simply use server-side apply to patch objects back to
desired state. If we could, this problem could be avoided entirely
because we'd never need to detect diffs (we'd just send our SSA patch to
the apiserver and have it do the work).

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>