ci(stale): update to run twice daily and remove failing print step"
Github rate limits their API and, since this action can do a large
number of actions in a short span of time, it defaults to 30 actions
in a single run. This is fine, but the only problem is that if several
issues are stale it will not get to them. As a result, this commit
updates the action to run twice daily which should help the issue.

Additionally, this commit removes a failing print step from the job.

Signed-off-by: Tyler Slaton <tyslaton@redhat.com>