fix: avoid bundle update errors during deletion
Our finalizer code is not integrated with our updater code, which
results in two multiple update calls being made during a single reconcile
when a bundle is deleted. The first call removes the finalizer, the
second attempts to update the status. The second call fails because the
bundle has been deleted in etcd after the finalizer is removed with the
first call.

This commit fully rips out the bundle updater and refactors things to
ensure there’s never more than one status and bundle update per
reconcile.