fix: update stale action to have more capacity to process items
Right now the bot is having a hard time processing our backlog due to its size. By default,
there are 30 operations allowed before it stops attempting to process things to avoid rate
limiting. We can get around this by increasing the amount of operations allowed per run.
This has more potential for rate limiting, but should be fine in off hours.

In addition to adding these changes, this commit sets the the action only run once per day
and updates the messaging it leaves when marking issues/prs as stale.