Consolidate bundle and BD controller code
We currently have a large amount of copy/pasted code in our various
in-tree provisioners. We have three bundle provisioners: plain,
registry, and helm. And we have two BD provisioners: plain and helm.

A very large majority of the code in each of these provisioners is
identical, and the only differences are in:
- how the bundle content is validated and converted before storing
- how the bundle content is validated and rendered into a helm chart
  before being applied to the cluster.

This PR introduces a few interfaces to capture these
implementation-specific differences along with generic bundle and BD
reconcilers that use these interfaces such that the copy/pasted code is
now shared and DRY-ed up.