Force install namespace for "other" objects
OLMv0 forces all objects in the bundle image to be created in the
install namespace even if left blank or filled with a different
namespace.

This commit forces all "other" objects to also have their namespace
populated via the converter, in the same way one would expect from
OLMv0.

This behaviour was tested using the cert-manager operator hub bundle in
version v1.12.2, where a Service without explicit namespace is part of
the bundle image.

Signed-off-by: Nico Schieder <nschieder@redhat.com>