Generic healtcheck
Adds a "generic" healthcheck function.

It returns true if all resources are healthy with nil error.
It returns false if any of the resources are not healthy, the error contains the GVK + resource name, and the error message of
each unhealthy resource.

The current list of supported resources is:
- Deployments
- StatefulSets
- DaemonSets
- ReplicaSets
- Pods
- APIServices
- CustomResourceDefinitions

If the resource is not supported, it is assumed to be healthy.

wip

wip

Signed-off-by: Joaquim Moreno Prusi <joaquim@redhat.com>